:lipstick: Move DelayedSequence to separate file
